+++ Robust.Client/Placement/PlacementManager.cs
@@ -573,7 +573,7 @@ namespace Robust.Client.Placement
 
         private void RequestPlacement(GridCoordinates coordinates)
         {
-            if (MapManager.GetGrid(coordinates.GridID).ParentMap.Index == MapId.Nullspace) return;
+            if (MapManager.GetGrid(coordinates.GridID).ParentMapId == MapId.Nullspace) return;
             if (CurrentPermission == null) return;
             if (!CurrentMode.IsValidPosition(coordinates)) return;
             if (Hijack != null && Hijack.HijackPlacementRequest(coordinates)) return;
