+++ Robust.Shared/Prototypes/EntityPrototype.cs
@@ -8,6 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Robust.Shared.Log;
+using Robust.Shared.Map;
 using Robust.Shared.Maths;
 using YamlDotNet.RepresentationModel;
 using Robust.Shared.Serialization;
